api updated

+++ src/api/posts.api.ts
@@ -1,15 +1,22 @@
 export async function httpCreatePost(title: string, content: string) {
   try {
+    const tokenCookie = document.cookie
+      .split("; ")
+      .find((row) => row.startsWith("token="));
+    if (!tokenCookie) {
+      throw new Error("Token not found in cookies");
+    }
+    const token = tokenCookie.split("=")[1];
     const res = await fetch(`${process.env.REACT_APP_API_URL}/post`, {
       method: "post",
       headers: {
         "content-type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify({
         title,
         content,
       }),
-      credentials: "include",
     });
     return {
       status: res.status,
@@ -26,16 +33,23 @@ export async function httpUpdatePost(
   content: string
 ) {
   try {
+    const tokenCookie = document.cookie
+      .split("; ")
+      .find((row) => row.startsWith("token="));
+    if (!tokenCookie) {
+      throw new Error("Token not found in cookies");
+    }
+    const token = tokenCookie.split("=")[1];
     const res = await fetch(`${process.env.REACT_APP_API_URL}/post/${postId}`, {
       method: "put",
       headers: {
         "content-type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify({
         title,
         content,
       }),
-      credentials: "include",
     });
     return {
       status: res.status,
@@ -66,17 +80,24 @@ export async function httpGetAllPosts() {
 
 export async function httpDeletePostById(postId: string) {
   try {
+    const tokenCookie = document.cookie
+      .split("; ")
+      .find((row) => row.startsWith("token="));
+    if (!tokenCookie) {
+      throw new Error("Token not found in cookies");
+    }
+    const token = tokenCookie.split("=")[1];
     const res = await fetch(`${process.env.REACT_APP_API_URL}/post/${postId}`, {
       method: "delete",
       headers: {
         "content-type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
-      credentials: "include",
     });
     return {
       status: res.status,
       data: await res.json(),
-    }
+    };
   } catch (error) {
     console.log(error);
   }
@@ -84,11 +105,20 @@ export async function httpDeletePostById(postId: string) {
 
 export async function httpGetPostsByUserId(userId: string) {
   try {
+    const tokenCookie = document.cookie
+      .split("; ")
+      .find((row) => row.startsWith("token="));
+    if (!tokenCookie) {
+      throw new Error("Token not found in cookies");
+    }
+    const token = tokenCookie.split("=")[1];
     const res = await fetch(
       `${process.env.REACT_APP_API_URL}/posts/${userId}`,
       {
         method: "get",
-        credentials: "include",
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
       }
     );
     return await res.json();
